fix: 'export pdf' file names and dependencies
