+++ package.json
@@ -36,6 +36,7 @@
     "@nestjs/swagger": "^7.4.2",
     "@prisma/client": "^6.5.0",
     "bcryptjs": "^3.0.2",
+    "express": "^5.1.0",
     "ioredis": "^5.6.0",
     "nodemailer": "^6.10.0",
     "passport-jwt": "^4.0.1",
